Moved custom widgets to seperate file
