+++ src/main_body.rs
@@ -3,14 +3,14 @@ use std::thread;
 use colored::Colorize;
 use eframe::{
     egui::{
-        self, CentralPanel, CursorIcon, Id, InnerResponse, LayerId, Layout,
-        Margin, Order, RichText, Sense, TextEdit, Ui,
+        self, CentralPanel, Id,  Layout,
+        Margin, RichText, TextEdit, 
     },
     emath::Align,
-    epaint::{self, Color32, Rect, Shape, Vec2},
+    epaint::{Color32, Vec2},
 };
 
-use crate::proxy::{Proxy, ProxyEvent};
+use crate::{proxy::{Proxy, ProxyEvent}, custom_widgets};
 
 pub fn main_body(proxy: &mut Proxy, ui: &mut egui::Ui) {
     let panel_frame = egui::Frame {
@@ -34,93 +34,6 @@ pub fn main_body(proxy: &mut Proxy, ui: &mut egui::Ui) {
                 logs_panel(proxy, ui);
             });
         });
-}
-
-// Toggle
-fn toggle_ui(ui: &mut egui::Ui, on: &mut bool) -> egui::Response {
-    let desired_size = ui.spacing().interact_size.y * egui::vec2(2.0, 1.0);
-    let (rect, mut response) = ui.allocate_exact_size(desired_size, egui::Sense::click());
-    if response.clicked() {
-        *on = !*on;
-        response.mark_changed();
-    }
-    response.widget_info(|| egui::WidgetInfo::selected(egui::WidgetType::Checkbox, *on, ""));
-
-    if ui.is_rect_visible(rect) {
-        let how_on = ui.ctx().animate_bool(response.id, *on);
-        let visuals = ui.style().interact_selectable(&response, *on);
-        let rect = rect.expand(visuals.expansion);
-        let radius = 0.5 * rect.height();
-        ui.painter()
-            .rect(rect, radius, visuals.bg_fill, visuals.bg_stroke);
-        let circle_x = egui::lerp((rect.left() + radius)..=(rect.right() - radius), how_on);
-        let center = egui::pos2(circle_x, rect.center().y);
-        ui.painter()
-            .circle(center, 0.75 * radius, visuals.bg_fill, visuals.fg_stroke);
-    }
-
-    response
-}
-
-// Drag
-pub fn drag_source(ui: &mut Ui, id: Id, body: impl FnOnce(&mut Ui)) {
-    let is_being_dragged = ui.memory(|mem| mem.is_being_dragged(id));
-
-    if !is_being_dragged {
-        let response = ui.scope(body).response;
-
-        let response = ui.interact(response.rect, id, Sense::drag());
-        if response.hovered() {
-            ui.ctx().set_cursor_icon(CursorIcon::Grab);
-        }
-    } else {
-        ui.ctx().set_cursor_icon(CursorIcon::Grabbing);
-
-        let layer_id = LayerId::new(Order::Tooltip, id);
-        let response = ui.with_layer_id(layer_id, body).response;
-
-        if let Some(pointer_pos) = ui.ctx().pointer_interact_pos() {
-            let delta = pointer_pos - response.rect.center();
-            ui.ctx().translate_layer(layer_id, delta);
-        }
-    }
-}
-
-// Drop
-pub fn drop_target<R>(ui: &mut Ui, body: impl FnOnce(&mut Ui) -> R) -> InnerResponse<R> {
-    let is_being_dragged = ui.memory(|mem| mem.is_anything_being_dragged());
-
-    let margin = Vec2::splat(0.);
-
-    let outer_rect_bounds = ui.available_rect_before_wrap();
-    let inner_rect = outer_rect_bounds.shrink2(margin);
-    let where_to_put_background = ui.painter().add(Shape::Noop);
-    let mut content_ui = ui.child_ui(inner_rect, *ui.layout());
-    let ret = body(&mut content_ui);
-    let outer_rect = Rect::from_min_max(outer_rect_bounds.min, content_ui.min_rect().max + margin);
-    let (rect, response) = ui.allocate_at_least(outer_rect.size(), Sense::hover());
-
-    let style = if is_being_dragged && response.hovered() {
-        ui.visuals().widgets.active
-    } else {
-        ui.visuals().widgets.inactive
-    };
-    let mut stroke = style.bg_stroke;
-    if is_being_dragged {
-        stroke.color = ui.visuals().gray_out(stroke.color);
-    }
-
-    ui.painter().set(
-        where_to_put_background,
-        epaint::RectShape {
-            rounding: style.rounding,
-            fill: ui.ctx().style().visuals.window_fill(),
-            stroke,
-            rect,
-        },
-    );
-
-    InnerResponse::new(ret, response)
 }
 
 // Run this on every frame to check if the port is valid
@@ -294,7 +207,7 @@ fn logs_panel(proxy: &mut Proxy, ui: &mut egui::Ui) {
                     if is_blocking {
                         ui.horizontal(|ui| {
                             ui.label("Deny Incoming");
-                            if toggle_ui(ui, &mut allow_requests_by_default).changed() {
+                            if custom_widgets::toggle_ui(ui, &mut allow_requests_by_default).changed() {
                                 proxy.switch_exclusion();
                             }
                             ui.label("Allow Incoming");
@@ -302,7 +215,7 @@ fn logs_panel(proxy: &mut Proxy, ui: &mut egui::Ui) {
 
                         egui::CollapsingHeader::new("Request Exclusion List").default_open(false).show_unindented(ui, |ui| {
                             ui.add_space(4.);
-                            let drop_response = drop_target(ui, |ui| {
+                            let drop_response = custom_widgets::drop_target(ui, |ui| {
                                 ui.group(|ui| {
                                     let exclusion_list = proxy.get_current_list();
                                     let num_rows = exclusion_list.len();
@@ -317,7 +230,7 @@ fn logs_panel(proxy: &mut Proxy, ui: &mut egui::Ui) {
                                                     // Truncate value so it fits better
                                                     let mut uri_truncated = uri.to_string();
                                                     if uri_truncated.len() > 35 {
-                                                        uri_truncated.truncate(35);
+                                                        uri_truncated.truncate(32);
                                                         uri_truncated += "...";
                                                     }
     
@@ -423,7 +336,7 @@ fn logs_panel(proxy: &mut Proxy, ui: &mut egui::Ui) {
                                                 ui.with_layout(
                                                     Layout::left_to_right(eframe::emath::Align::Center),
                                                     |ui| {
-                                                        drag_source(ui, item_id, |ui| {
+                                                        custom_widgets::drag_source(ui, item_id, |ui| {
                                                             ui.horizontal(|ui| {
                                                                 ui.with_layout(
                                                                     Layout::left_to_right(Align::Max),
